feat(hub): export shared WrappedDeksBlob primitive (#44)
Adds packages/hub/src/team/wrapped-deks.ts — the canonical
implementation of "serialize a DEK set, encrypt it under a
credential-derived AES-GCM key." Both mintPaperRecoveryEntry (in
team/recovery.ts) and enrollPasswordAuthenticator (in
@noy-db/on-password) now delegate to this single helper, instead of
reimplementing the same crypto inline.

## API

  interface WrappedDeksBlob {
    salt: string         // base64 PBKDF2 salt
    iv: string           // base64 AES-GCM IV
    wrappedDeks: string  // base64 ciphertext of {deks: {coll: rawB64}}
  }

  mintWrappedDeksBlob(deks, credential): Promise<WrappedDeksBlob>
  unwrapDeksFromBlob(blob, credential): Promise<Map<string, CryptoKey>>

## Consumer composition

  - PaperRecoveryEntry now extends WrappedDeksBlob — the three crypto
    fields (salt, iv, wrappedDeks) come from the shared primitive;
    codeId + enrolledAt are paper-recovery's own metadata.
  - mintPaperRecoveryEntry is now a 4-line wrapper that calls
    mintWrappedDeksBlob and adds {codeId, enrolledAt}.
  - unwrapDeksFromPaperEntry is a 1-line wrapper around
    unwrapDeksFromBlob.
  - enrollPasswordAuthenticator delegates to mintWrappedDeksBlob;
    the slot envelope still stores `salt` inside `meta` for
    backward compatibility with pre-#44 slot format (the issue
    explicitly defers moving it to top-level for parity).
  - unwrapDeksWithPassword reconstructs a WrappedDeksBlob from the
    slot's split layout (salt in meta, iv + wrappedDeks at top
    level) and delegates to unwrapDeksFromBlob.

## Wire format

UNCHANGED. Both the _meta/recovery-paper document shape and the
KeyringAuthenticator wrap-DEKs slot format are byte-identical to
pre-#44. The refactor is purely code-level deduplication; existing
on-disk entries continue to round-trip.

## Dead code removed

  - team/recovery.ts: deriveRecoveryWrappingKey, bytesToBase64,
    base64ToBytes (~30 LOC)
  - on-password/src/index.ts: derivePasswordWrappingKey,
    bytesToBase64, base64ToBytes, SALT_BYTES, IV_BYTES, subtle
    constants (~40 LOC)

## Verification

  - 2406 / 2410 full-repo tests pass (4 skipped IDB-perf benches)
  - Typecheck: 107 / 107 targets clean
  - Lint: 106 / 106 packages clean

## Forward-compat note

Future on-* packages that wrap DEKs (per the auth-landscape
decision rules) reach for `mintWrappedDeksBlob` directly — no
reimplementation needed. The slot-format consolidation (moving
password's `salt` from `meta` to top-level for parity with
PaperRecoveryEntry) remains deferred as a separate breaking change.

Closes #44
